// src/bindings/models/TodoSyncOperation.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { SyncOperationKind } from "./SyncOperationKind";
import type { TodoPatch } from "./TodoPatch";

export type TodoSyncOperation = { operationId: string, todoId: string, kind: SyncOperationKind, occurredAt: string, patch?: TodoPatch, };